+++ test/mssql_test.cpp
@@ -221,6 +221,11 @@ TEST_CASE_METHOD(mssql_fixture, "test_batch_insert_integral", "[mssql][batch][in
     test_batch_insert_integral();
 }
 
+TEST_CASE_METHOD(mssql_fixture, "test_batch_insert_null", "[mssql][batch][null]")
+{
+    test_batch_insert_null();
+}
+
 TEST_CASE_METHOD(mssql_fixture, "test_batch_insert_string", "[mssql][batch][string]")
 {
     test_batch_insert_string();
@@ -1060,6 +1065,16 @@ PRIMARY KEY(t2_fid)
 TEST_CASE_METHOD(mssql_fixture, "test_integral", "[mssql][integral]")
 {
     test_integral<mssql_fixture>();
+}
+
+TEST_CASE_METHOD(mssql_fixture, "test_integral_small_types", "[mssql][integral]")
+{
+    test_integral_small_types();
+}
+
+TEST_CASE_METHOD(mssql_fixture, "test_integral_small_types_batch", "[mssql][integral][batch]")
+{
+    test_integral_small_types_batch();
 }
 
 TEST_CASE_METHOD(mssql_fixture, "test_move", "[mssql][move]")